flatten app configs
