+++ packages/app/src/Router.tsx
@@ -7,7 +7,7 @@ import { StakingApi } from 'StakingApi'
 import { useActiveAccount } from '@polkadot-cloud/connect'
 import { useEffectIgnoreInitial } from '@w3ux/hooks'
 import { extractUrlValue } from '@w3ux/utils'
-import { PagesConfig } from 'config/pages'
+import { PagesConfig } from 'config'
 import { getUnixTime } from 'date-fns'
 import {
 	onConversionEvent,
