--- packages/app/src/config/help.ts
@@ -1,24 +1,0 @@
-// Copyright 2026 @polkadot-cloud/polkadot-staking-dashboard authors & contributors
-// SPDX-License-Identifier: GPL-3.0-only
-
-// Help items that are not documented on staking docs
-export const HelpNoDocs = [
-	'bondedInPool',
-	'decentralizationAnalyticsPeriod',
-	'geolocationOfEachNomination',
-	'ledgerAppNotOnLatestRuntimeVersion',
-	'ledgerHardwareWallets',
-	'ledgerRejectedTransaction',
-	'ledgerRequestTimeout',
-	'minimumToCreatePool',
-	'minimumToJoinPool',
-	'nominationPayoutDistribution',
-	'openAppOnLedger',
-	'poolRewardHistory',
-	'rewardsByCountryAndNetwork',
-	'selfStake',
-	'totalPayoutsAnalysed',
-	'validatorRewardHistory',
-	'yourBalance',
-	'halving',
-]
